vector-store: rename integration test from usearch to vs_index
this will allow for testing multiple engines (diskann, usearch) from a single unified test file
