+++ crates/vector-store/tests/integration/mtls.rs
@@ -11,7 +11,7 @@ use crate::tls_utils::generate_client_identity;
 use crate::tls_utils::generate_server_cert;
 use crate::tls_utils::init;
 use crate::tls_utils::read_cert;
-use crate::usearch::test_config;
+use crate::vs_index::test_config;
 use crate::wait_for;
 use tempfile::NamedTempFile;
 use tokio::sync::mpsc::Sender;
