+++ crates/vector-store/tests/integration/vs_index.rs
